% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{display_all_palettes}
\alias{display_all_palettes}
\title{Display all available palettes}
\usage{
display_all_palettes(n = NULL)
}
\arguments{
\item{n}{Number of colors to display for each palette}
}
\value{
A plot (invisible NULL)
}
\description{
Creates a multi-panel plot showing all available palettes
}
\examples{
display_all_palettes()
display_all_palettes(10)
}
