% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{display_palette}
\alias{display_palette}
\title{Display a color palette}
\usage{
display_palette(name, n = NULL)
}
\arguments{
\item{name}{Character name of the palette}

\item{n}{Number of colors to display}
}
\value{
A plot (invisible NULL)
}
\description{
Creates a simple plot displaying the colors in a palette
}
\examples{
display_palette("sunset")
display_palette("ocean", 10)
}
