% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_palette}
\alias{get_palette}
\title{Get a color palette}
\usage{
get_palette(name, n = NULL, reverse = FALSE)
}
\arguments{
\item{name}{Character name of the palette}

\item{n}{Number of colors desired. If provided, the palette will be interpolated
to produce the desired number of colors.}

\item{reverse}{Logical. Should the palette be reversed?}
}
\value{
A character vector of colors
}
\description{
Returns a vector of colors from a specified palette
}
\examples{
get_palette("sunset", 3)
get_palette("ocean", 10)
get_palette("forest", 5, reverse = TRUE)
}
