% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_color}
\alias{palette_color}
\title{Apply palette to a plot}
\usage{
palette_color(name, n = NULL, reverse = FALSE)
}
\arguments{
\item{name}{Character name of the palette}

\item{n}{Number of colors needed (typically matches the number of groups)}

\item{reverse}{Logical. Should the palette be reversed?}
}
\value{
A function that can be used with the col parameter in plotting functions
}
\description{
Helper function to quickly apply a palette to common plotting functions
}
\examples{
\dontrun{
# For base R plots
plot(mpg ~ disp, data = mtcars, col = palette_color("sunset")(factor(mtcars$cyl)))

# For ggplot2 (after setting up scale_color_palette or scale_fill_palette)
library(ggplot2)
ggplot(mtcars, aes(mpg, disp, color = factor(cyl))) +
  geom_point() +
  scale_color_manual(values = palette_color("sunset")(3))
}
}
