% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Available color palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
palettes
}
\description{
A list containing all available color palettes
}
\keyword{datasets}
