% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{qipalette-package}
\alias{qipalette-package}
\title{Color Palettes}
\description{
A collection of color palettes for data visualization
}
\examples{
# View all palettes
display_all_palettes()

# Use a specific palette
pal <- get_palette("sunset", n = 5)
plot(1:5, 1:5, col = pal, pch = 19, cex = 3)

}
