% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qipalette.R
\docType{package}
\name{qipalette}
\alias{qipalette}
\title{Color Palettes for Data Visualization}
\description{
A collection of visually appealing color palettes for data visualization in R.
This package provides functions to easily access and use these palettes in
various plotting systems.
}
\section{Main functions}{

\itemize{
\item \code{\link{get_palette}} - Get a color palette by name
\item \code{\link{display_palette}} - Display a specific color palette
\item \code{\link{display_all_palettes}} - Display all available palettes
\item \code{\link{palette_color}} - Helper for applying palettes to plots
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quadram-institute-bioscience/qipalette}
  \item Report bugs at \url{https://github.com/quadram-institute-bioscience/qipalette/issues}
}

}
\author{
\strong{Maintainer}: Andrea Telatin \email{andrea.telatin@quadram.ac.uk}

}
