% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale_color_palette}
\alias{scale_color_palette}
\title{Scale color for ggplot2}
\usage{
scale_color_palette(palette = "sunset", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of the palette}

\item{discrete}{Logical. Is the scale discrete or continuous?}

\item{reverse}{Logical. Should the palette be reversed?}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn()}
}
\value{
A ggplot2 scale
}
\description{
A color scale for ggplot2 based on the package palettes
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point(size = 4) +
  scale_color_palette("sunset")
}
}
