% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale_fill_palette}
\alias{scale_fill_palette}
\title{Scale fill for ggplot2}
\usage{
scale_fill_palette(palette = "sunset", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of the palette}

\item{discrete}{Logical. Is the scale discrete or continuous?}

\item{reverse}{Logical. Should the palette be reversed?}

\item{...}{Additional arguments passed to discrete_scale() or scale_fill_gradientn()}
}
\value{
A ggplot2 scale
}
\description{
A fill scale for ggplot2 based on the package palettes
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  scale_fill_palette("ocean", guide = "none")
}
}
